Path following car AI using only force and torque
Successful pathfinding package integration and use.
Huge ground vehicle AI control script.
Vehicles are actually rigid bodies, so we can only apply forces and torques to it.
Wheel colliders allow to simulate wheels.
However, I wanted to try to do that without wheel colliders, as I thought they might be hard on performance (in a project where it should be managed carefully).
Another big reason was my suspiscion of wheel colliders being worth it only with an AI way smarter than I could have taken on, to decide on when to reverse etc for instance
That's why I tried to simulate car turns etc using only forces and torques.
It sometimes shows surprisingly good results, but is nearly always possible to find a weird situation where it becomes unstable.
It also uses a lot of hand-picked parameters which would have to be corrected every time a new vehicle would be imagined.
These parameters are quite intricate and are often intertwined. On top of that, there are 3 PID controllers used.

I am kind of satisfied with this somwhat working result, despite the challenge - the turns and reverses are fairly convincing.
The central downside is the really narrow turns with a way too high angular velocity.

Beucase of all those parameters and calculations, I am actually really unsure about this being lighter on performance than just using wheel colliders.

Still satisfied with the work. I'll try with wheel colliders now. I am confident it will lead to something really cool, as only the "motion command" part will change a bit, but all of the other already working logic is almost sure to be totally compatible.