Path following car AI using wheel colliders
Basically the same system as the attempt with only forces and torque, but a bit cleaner and improved.
Path-relative information etc encapsulated into a InternalPath struct in the GVWheelsAIController class, for instance.

The good news is that it's really quite accurate at following paths and reversing appropriately and correctly !
It can also go backwards and reverse to that correctly !

However, orientation once at the end of the path doesn't work and has been deleted because from here as it was causing a mess.
The current code is simplified, without those attempts, as it was complicating the code and the approach wasn't quite good enough.
Moreover, it is still unsure whether this should even be a feature, and even if it would be, how it should play out.
So I just scrapped it. It works really well without it.
And if needed, this part of the code could always be rewritten at later stages.