Great start for working pathtracking + formations
Wheeled vehicle pathtracking is working really well !!
There is still a number of "hardcoded" things, but in a few cases, I'd say it's acceptable or even maybe a good thing.
It will need cleaning and optimisation though, as it's still kind of dirty, becuase of "leftover" commented code, for instance.
It features :
- path tracking
- reversing
- path tracking backwards
- aligning to a desired orientation at the end of the path
The whole thing uses "wheel" physics etc, combined with steering behaviors to determine "desired location"
For now, the only implemented "steering behavior" is the path tracking, which takes cares of all of the features mentioned above.
Whether it should be broken down into smaller behavior will be adressed later, as there is no point in breaking it now when it's working, and mostly because having "one behavior" taking care of all that makes logical sense.
Later, other behaviors may take care of local avoidance

Also, the "path tracking" behavior could be done in a (slightly?) different approach.
The current relies on rather well smoothed paths (but can totally adapt well with quite natural reversing !), regularly and rather closely sampled waypoints in the path, followed one after each other.
But the system would still work great (except maybe for "transition phases" when reversing or going backwards or getting to a specified orientation at destination, which would then need a liiitle more patching) with another approach.
This approach would consist in the "target" point we steer towards not being a waypoint in the path, but the intersection point between a circle of a certain radius around the agent, and the path itself.
I think this would really really be worth exploring and trying out.

Formation movement basics have also been laid down, and it's actually not that bad.
For now, it basically makes a path from offsetted waypoints around a reference unit's path for the others in the formation to follow.
The offsetting is the problematic part, because it is quite imprecise.
But this method allows to calculate a path only once for all the agents in the formation, this is a critical issue which will have to be solved.

But I think that another method which has my confidence should also be tried out.
With it, agents would steer to a their dynamically offsetted formation positions, except when they don't have "line of sight" to it. Then, there should be a quick repath to the closest position in the path where free line of sight is expected.
My intuition (without testing backing) tells me this should be a bit more performance heavy (but only by a unsignificant fraction (I hope lol)) but possible more robust and natural.

Also, in any case, there will have to be some way of getting everyone "to the same progress" in the path. Like by slowing down those that are in advance for instance.
We can easily think of that being done based on comparisons of for example the speed or "path completion percentage" or "distance from desired offsetted position" of an agent to that of the other agents, with a "general evaluation" of them (could be as simple as mean)

All of this stuff would benefit from further optimisation in the form of better-suited and more efficient data representation structures and/or algorithms.
But for now, the Astar Pathfinding Project is great to work with, and allows for high end features to be incorporated / tested / used quickly, efficiently, and in a really clean and clear way.
Possibly, hierarchical (cooperative?) theta star hybrid algorithm, with quad tree partitioning or something like that on the (rather still far away horizon) would be great, and would not break everything at all. (nice system architecture 8D)

Finally a few things concerning order-giving (quite ok if I recall) and execution have been prepared (and that even includes an htn planning framework, but no attempt to make use of it yet.)

Whew ! Actual cool stuff here !